requirements.txt
pytest
moto
pytest-mock
requests
python-dotenv
openai
Flask
firebase-admin